Create `WorkflowFixtureNotAvailable` exception.
Thrown when a workflow fixture cannot be instantiated
due to no definition existing for it.
