+++ virtool_workflow/__init__.py
@@ -1,7 +1,7 @@
 """An SDK for defining Virtool workflows."""
 from .workflow import Workflow, WorkflowStep
 from .context import WorkflowExecutionContext
-from .workflow_fixture import fixture, WorkflowFixture
+from virtool_workflow.fixtures.workflow_fixture import fixture, WorkflowFixture
 
 __fixtures__ = [
     "virtool_workflow.storage.paths"
