+++ virtool_workflow/analysis/read_paths.py
@@ -10,7 +10,7 @@ from virtool_workflow.analysis.analysis_info import AnalysisArguments
 from virtool_workflow.analysis.cache import fetch_cache, create_cache
 from virtool_workflow.execute import FunctionExecutor
 from virtool_workflow.storage.utils import copy_paths
-from virtool_workflow.workflow_fixture import WorkflowFixtureScope
+from virtool_workflow.fixtures.scope import WorkflowFixtureScope
 from virtool_workflow_runtime.db import VirtoolDatabase
