+++ virtool_workflow/fixtures/__init__.py
@@ -1,0 +1,1 @@
+"""pytest style fixtures for use in Virtool Workflows."""
